fix: trigger resolution on dirty props
Trigger resolution whenever a prop is dirty. For resource-related props, like
Colors and Files, this will result in attempting to fetch the resource whenever
the prop value changes (from the cache or network).
