+++ .changeset/tiny-pots-know.md
@@ -1,0 +1,5 @@
+---
+"@makeswift/runtime": patch
+---
+
+fix: trigger resolution whenever a prop changes, resolving an issue where resources may not appear when a prop changes.
